refactor: extract SunCycleCard into separate file
