+++ Localizable.xcstrings
@@ -296,6 +296,7 @@
       }
     },
     "Daylight" : {
+      "extractionState" : "stale",
       "localizations" : {
         "de" : {
           "stringUnit" : {
